# Add benchmark to track Julia's array allocation performance relative to Python