--- a/test/perf/arrayalloc/Makefile
+++ b/test/perf/arrayalloc/Makefile
@@ -1,14 +1,14 @@
-ifndef JULIABIN
-	JULIABIN = julia
-endif
+JULIAHOME = $(abspath ../../..)
+include $(JULIAHOME)/Make.inc
+
 ifndef PYTHONBIN
-	PYTHONBIN = python3
+	PYTHONBIN = python
 endif
 
 python:
 	$(PYTHONBIN) lucompletepiv.py
 
 julia:
-	$(JULIABIN) lucompletepiv.jl
+	@$(call spawn,$(JULIA_EXECUTABLE)) lucompletepiv.jl
 
 all: python julia