--- a/test/perf/arrayalloc/Makefile
+++ b/test/perf/arrayalloc/Makefile
@@ -1,0 +1,14 @@
+ifndef JULIABIN
+	JULIABIN = julia
+endif
+ifndef PYTHONBIN
+	PYTHONBIN = python3
+endif
+
+python:
+	$(PYTHONBIN) lucompletepiv.py
+
+julia:
+	$(JULIABIN) lucompletepiv.jl
+
+all: python julia