Although Julia is often very fast, it is less fast when the code is written in a traditional MATLAB/Numpy style with a lot of temporary arrays. It was recently experienced in a linear algebra benchmark suggested by @jiahao for which a Numpy implementation was quite a bit faster than Julia.

This PR adds the benchmark to then benchmark suite such that it is easier to track improvements in our array allocation/reuse speed. The benchmarks are written for Python 3 because it appeared to be faster the 2 on my machine. In the timings below, I use latest Julia master.

It should be noticed that the new GC has already helped a lot.

```sh
python3 lucompletepiv.py
size  100 matrix factorized in  0.005 seconds
size  250 matrix factorized in  0.038 seconds
size  500 matrix factorized in  0.218 seconds
size 1000 matrix factorized in  1.713 seconds

julia-dev lucompletepiv.jl
Julia version with copy slices
size  100 matrix factorized in  0.015 seconds
size  250 matrix factorized in  0.104 seconds
size  500 matrix factorized in  0.538 seconds
size 1000 matrix factorized in  4.027 seconds

Julia version with view slices
size  100 matrix factorized in  0.004 seconds
size  250 matrix factorized in  0.079 seconds
size  500 matrix factorized in  0.442 seconds
size 1000 matrix factorized in  3.632 seconds
```

cc: @ViralBShah