@timholy 2015-05-26
I'm not certain this is entirely a question of allocation. Profiling seems to suggest that it spends quite a lot of time inside loops, specifically in these places:
https://github.com/JuliaLang/julia/blob/96cc97065ef07a741e21dce3d47768d632ae131a/base/operators.jl#L361
https://github.com/JuliaLang/julia/blob/96cc97065ef07a741e21dce3d47768d632ae131a/base/array.jl#L1206-L1209
https://github.com/JuliaLang/julia/blob/96cc97065ef07a741e21dce3d47768d632ae131a/base/array.jl#L805-L807
https://github.com/JuliaLang/julia/blob/96cc97065ef07a741e21dce3d47768d632ae131a/base/multidimensional.jl#L313-L316

I wouldn't be entirely surprised if the performance were improved by adding `@inbounds` and `@simd` in more places.