@andreasnoack 2015-05-26
I tried to replace elementwise `abs` with a non-comprehension version with `@inbounds` and `@simd` and it gave around 0.4 seconds speedup for 1000x1000 matrices. I haven't been able to get a speedup for the `-` method.

I compared the profiling output to similar output in Python/Numpy and even with the optimized version of `abs`, Julia spends way more time at that line.