@timholy 2015-05-26
So basically, we have a slow `abs`?

For some reason, adding `@simd` doesn't actually cause the `-` loop to SIMDify. CC @ArchRobison.
```jl
function myminus3{S,T}(A::AbstractArray{S}, B::AbstractArray{T})
    F = similar(A, promote_type(S,T), promote_shape(size(A),size(B)))
    @simd for i in eachindex(A,B)
        @inbounds F[i] = A[i] - B[i]
    end
    return F
end
A = rand(Float32,1000,1000);
S = sub(A, 1:999, 1:999);
@code_llvm myminus3(S,S)
```
(Contrast `@code_llvm myminus3(A,A)`, which does achieve vectorization.)