@ViralBShah 2015-05-26
I have been observing this as well. We need to take a call on array views - which is probably the main issue at hand here, and always better GC.

cc: @jakebolewski