@timholy 2015-05-26
I'm not quite sure I understand what you're saying, but everything is inlined in the SubArray case: `@code_typed` shows no `call` statements except those used in creating `F`.

Interestingly, it has nothing to do with the `CartesianIndex` iteration or `eachindex`, as
```jl
function myminus4{S,T}(A::AbstractArray{S}, B::AbstractArray{T})
    F = similar(A, promote_type(S,T), promote_shape(size(A),size(B)))
    for j = 1:size(A,2)
        @simd for i = 1:size(A,1)
            @inbounds F[i,j] = A[i,j] - B[i,j]
        end
    end
    return F
end
```
is also not SIMD-vectorized for `S` but is for `A`. Interestingly, the simpler
```jl
function mysum{T}(A::AbstractArray{T})
    s = 0.0
    for j = 1:size(A,2)
        @simd for i = 1:size(A,1)
            @inbounds s += A[i,j]
        end
    end
    return s
end
```
does vectorize for both `A` and `S`. But change that to
```jl
function mysum!{T}(F, A::AbstractArray{T})
    for j = 1:size(A,2)
        @simd for i = 1:size(A,1)
            @inbounds F[i,j] += A[i,j]
        end
    end
    return F
end
```
and it fails (for `F::Array{Float32,2}` and `A::SubArray{Float32,2,Array{Float32,2},Tuple{UnitRange{Int64},UnitRange{Int64}},1}`).