@timholy 2015-05-26
Regarding `abs`, comparison with `matlab -singleCompThread`:
```
>> A = randn(10^7,1);
>> tic; B = abs(A); toc
Elapsed time is 0.060322 seconds.
>> tic; B = abs(A); toc
Elapsed time is 0.072021 seconds.
```
```jl
julia> A = randn(10^7);

julia> @time B = abs(A);
  80.982 milliseconds (4070 allocations: 78319 KB, 3.79% gc time)

julia> @time B = abs(A);
  76.090 milliseconds (6 allocations: 78125 KB, 24.01% gc time)

julia> @time B = abs(A);
  65.412 milliseconds (6 allocations: 78125 KB, 77.06% gc time)
```
so julia and matlab seem no different. Moreover,
```jl
julia> function myabs(A)
           B = similar(A)
           @simd for i in eachindex(A)
               @inbounds B[i] = abs(A[i])
           end
           B
       end
myabs (generic function with 1 method)

julia> @time B = myabs(A);
  62.440 milliseconds (4298 allocations: 78337 KB, 20.63% gc time)

julia> @time B = myabs(A);
  39.526 milliseconds (6 allocations: 78125 KB, 5.20% gc time)
```
is considerably faster.