@timholy 2015-05-26
Likewise, python:
```
>>> A = randn(10000000,1)
>>> size(A)
10000000
>>> start_time = timeit.default_timer(); B = abs(A); elapsed = timeit.default_timer() - start_time; print elapsed
0.0585150718689
>>> start_time = timeit.default_timer(); B = abs(A); elapsed = timeit.default_timer() - start_time; print elapsed
0.126361846924
>>> start_time = timeit.default_timer(); B = abs(A); elapsed = timeit.default_timer() - start_time; print elapsed
0.0634398460388
```
@andreasnoack, it would be helpful to quantify "way more time." Is it so different that it suggests this loop is somehow being elided in python?